camel-16400: adjust integration tests to avoid component plugin from recreating them for camel-braintree
