+++ components/camel-braintree/src/test/java/org/apache/camel/component/braintree/DiscountGatewayIT.java
@@ -14,14 +14,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.braintree.integration;
+package org.apache.camel.component.braintree;
 
 import java.util.List;
 
-import com.braintreegateway.Plan;
+import com.braintreegateway.Discount;
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.component.braintree.AbstractBraintreeTestSupport;
-import org.apache.camel.component.braintree.internal.PlanGatewayApiMethod;
+import org.apache.camel.component.braintree.internal.BraintreeApiCollection;
+import org.apache.camel.component.braintree.internal.DiscountGatewayApiMethod;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
@@ -31,15 +31,16 @@ import org.slf4j.LoggerFactory;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 @EnabledIfSystemProperty(named = "braintreeAuthenticationType", matches = ".*")
-public class PlanGatewayIT extends AbstractBraintreeTestSupport {
+public class DiscountGatewayIT extends AbstractBraintreeTestSupport {
 
-    private static final Logger LOG = LoggerFactory.getLogger(PlanGatewayIT.class);
-    private static final String PATH_PREFIX = getApiNameAsString(PlanGatewayApiMethod.class);
+    private static final Logger LOG = LoggerFactory.getLogger(DiscountGatewayIT.class);
+    private static final String PATH_PREFIX
+            = BraintreeApiCollection.getCollection().getApiName(DiscountGatewayApiMethod.class).getName();
 
     @Disabled
     @Test
     public void testAll() throws Exception {
-        final List<Plan> result = requestBody("direct://ALL", null, List.class);
+        final List<Discount> result = requestBody("direct://ALL", null, List.class);
 
         assertNotNull(result, "all result");
         LOG.debug("all: " + result);
