+++ components/camel-braintree/src/test/java/org/apache/camel/component/braintree/PaymentMethodNonceGatewayIT.java
@@ -14,14 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.braintree.integration;
+package org.apache.camel.component.braintree;
 
-import com.braintreegateway.CreditCardVerification;
-import com.braintreegateway.ResourceCollection;
 import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.component.braintree.AbstractBraintreeTestSupport;
 import org.apache.camel.component.braintree.internal.BraintreeApiCollection;
-import org.apache.camel.component.braintree.internal.CreditCardVerificationGatewayApiMethod;
+import org.apache.camel.component.braintree.internal.PaymentMethodNonceGatewayApiMethod;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
@@ -31,45 +28,44 @@ import org.slf4j.LoggerFactory;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 @EnabledIfSystemProperty(named = "braintreeAuthenticationType", matches = ".*")
-public class CreditCardVerificationGatewayIT extends AbstractBraintreeTestSupport {
+public class PaymentMethodNonceGatewayIT extends AbstractBraintreeTestSupport {
 
-    private static final Logger LOG = LoggerFactory.getLogger(CreditCardVerificationGatewayIT.class);
+    private static final Logger LOG = LoggerFactory.getLogger(PaymentMethodNonceGatewayIT.class);
     private static final String PATH_PREFIX
-            = BraintreeApiCollection.getCollection().getApiName(CreditCardVerificationGatewayApiMethod.class).getName();
+            = BraintreeApiCollection.getCollection().getApiName(PaymentMethodNonceGatewayApiMethod.class).getName();
+
+    // TODO provide parameter values for create
+    @Disabled
+    @Test
+    public void testCreate() throws Exception {
+        // using String message body for single parameter "paymentMethodToken"
+        final com.braintreegateway.Result result = requestBody("direct://CREATE", null);
+
+        assertNotNull(result, "create result");
+        LOG.debug("create: " + result);
+    }
 
     // TODO provide parameter values for find
     @Disabled
     @Test
     public void testFind() throws Exception {
-        // using String message body for single parameter "id"
-        final CreditCardVerification result = requestBody("direct://FIND", null, CreditCardVerification.class);
+        // using String message body for single parameter "paymentMethodNonce"
+        final com.braintreegateway.PaymentMethodNonce result = requestBody("direct://FIND", null);
 
         assertNotNull(result, "find result");
         LOG.debug("find: " + result);
-    }
-
-    // TODO provide parameter values for search
-    @Disabled
-    @Test
-    public void testSearch() throws Exception {
-        // using com.braintreegateway.CreditCardVerificationSearchRequest message body for single parameter "query"
-        final ResourceCollection<CreditCardVerification> result
-                = requestBody("direct://SEARCH", null, ResourceCollection.class);
-
-        assertNotNull(result, "search result");
-        LOG.debug("search: " + result);
     }
 
     @Override
     protected RouteBuilder createRouteBuilder() throws Exception {
         return new RouteBuilder() {
             public void configure() {
+                // test route for create
+                from("direct://CREATE")
+                        .to("braintree://" + PATH_PREFIX + "/create?inBody=paymentMethodToken");
                 // test route for find
                 from("direct://FIND")
-                        .to("braintree://" + PATH_PREFIX + "/find?inBody=id");
-                // test route for search
-                from("direct://SEARCH")
-                        .to("braintree://" + PATH_PREFIX + "/search?inBody=query");
+                        .to("braintree://" + PATH_PREFIX + "/find?inBody=paymentMethodNonce");
             }
         };
     }
